RCT stats/info improvements
* Now accepts both mentions and case insensitive strings
* Shows player's placement on the ladder
* Displays player's Custom Account Icon if they own one
* Embed timestamp is now the timestamp of last info retrieval
* Option to preserve the message and prevent its scheduled deletion

+++ extensions/stats.py
@@ -2,11 +2,13 @@ import asyncio
 import timeit
 import collections
 
+import aiohttp
 import discord
 from discord.ext import commands
 
 import config
-from extensions.checks import database_ready
+from extensions.checks import database_ready, is_senior
+from extensions.masterserver import nick2id
 
 # TO DO: timeout wait_for reaction
 
@@ -15,21 +17,44 @@ class Stats(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    # TO DO: clean up branching, async functions
+    # TO DO: clean up branching, member to user, CAI
     @commands.command(aliases=["info", "sheet", "rank"])
     @database_ready()
-    async def stats(self, ctx, member: discord.Member = None):
-        """Gets user's RCT game info from the sheet"""
+    async def stats(self, ctx, member: str = ""):
+        """Gets user's RCT game info from the sheet."""
         start = timeit.default_timer()
+        list_of_lists = config.LIST_OF_LISTS
         check_failed = False
 
-        if member is None:
-            member = ctx.message.author
-        member_discord_id = str(member.id)
-        requester_discord_id = str(ctx.message.author.id)
+        # Remove Discord ID dep at some point. This whole branching requires a better solution but keeping it for now.
+        if member == "":
+            member = ctx.author
+            member_discord_id = str(ctx.author.id)
+        elif len(ctx.message.raw_mentions) > 0:
+            member = discord.utils.find(
+                lambda m: m.id == ctx.message.raw_mentions[0], ctx.guild.members
+            )
+            member_discord_id = str(ctx.message.raw_mentions[0])
+        else:
+            found_em = False
+            for row in list_of_lists:
+                if row[1].lower() == member.lower():
+                    if row[32].isdigit():
+                        lu_member_id = int(row[32])
+                        found_em = True
+                        break
+            if not found_em:
+                return await ctx.send("That player has never joined Discord.")
+            member = discord.utils.find(
+                lambda m: m.id == lu_member_id, ctx.guild.members
+            )
+            member_discord_id = str(lu_member_id)
+        requester_discord_id = str(ctx.author.id)
         requester_name = None
+        # member = ctx.guild.get_member(int(member_discord_id))
 
-        for row in config.LIST_OF_LISTS:
+        # This needs to be split. Keeping for now.
+        for row in list_of_lists:
             if row[32] == member_discord_id:
                 row_values = row
             if row[32] == requester_discord_id:
@@ -38,7 +63,7 @@ class Stats(commands.Cog):
         nick = row_values[1]
         nick_lower = nick.lower()
 
-        # check trivia spreadsheet for points
+        # Check trivia spreadsheet for points.
         try:
             for row in config.LIST_OF_LISTS_TRIVIA:
                 if row[0].lower() == nick_lower:
@@ -66,7 +91,7 @@ class Stats(commands.Cog):
         else:
             bonus = ", ".join(bonus) + " games"
 
-        # this cycle
+        # This cycle.
         seconds = int(row_values[3])
         dhms = ""
         for scale in 86400, 3600, 60:
@@ -75,18 +100,9 @@ class Stats(commands.Cog):
                 dhms += "{0:02d}:".format(result)
         dhms += "{0:02d}".format(seconds)
 
-        # if seconds>=60:
-        #     dhms = dhms.split(':')
-        #     gametime = "{0} minutes {1} seconds".format(dhms[0],dhms[1])
-        # elif seconds>=3600:
-        #     dhms = dhms.split(':')
-        #     gametime = "{0} hours {1} minutes {2} seconds".format(dhms[0],dhms[1],dhms[2])
-        # else:
-        #     gametime = "{0} seconds".format(dhms)
+        gametime = f"{dhms}"
 
-        gametime = "{0}".format(dhms)
-
-        # total
+        # Total.
         seconds_total = int(row_values[6])
         dhms_total = ""
         for scale_total in 86400, 3600, 60:
@@ -95,19 +111,7 @@ class Stats(commands.Cog):
                 dhms_total += "{0:02d}:".format(result_total)
         dhms_total += "{0:02d}".format(seconds_total)
 
-        # if seconds_total>=60:
-        #     dhms_total = dhms_total.split(':')
-        #     gametime_total = "{0} minutes {1} seconds".format(dhms_total[0],dhms_total[1])
-        # elif seconds_total>=3600:
-        #     dhms_total = dhms_total.split(':')
-        #     gametime_total = "{0} hours {1} minutes {2} seconds".format(dhms_total[0],dhms_total[1],dhms_total[2])
-        # elif seconds_total>=86400:
-        #     dhms_total = dhms_total.split(':')
-        #     gametime_total = "{0} days {1} hours {2} minutes {3} seconds".format(dhms_total[0],dhms_total[1],dhms_total[2],dhms_total[2])
-        # else:
-        #     gametime_total = "{0} seconds".format(dhms_total)
-
-        gametime_total = "{0}".format(dhms_total)
+        gametime_total = f"{dhms_total}"
 
         heroes = []
         players = []
@@ -135,8 +139,6 @@ class Stats(commands.Cog):
             for i in range(0, len(players)):
                 if players[i].lower() == nick_lower:
                     heroes_played.append(heroes[i])
-                    # playerName=players[i]
-            # games_played = collections.Counter(heroes_played)
             games_played = str(len(heroes_played))
 
         def is_current_member():
@@ -149,84 +151,205 @@ class Stats(commands.Cog):
             )  # true, 1 legacy
 
         current_member = is_current_member()
-        if current_member:
-            former = ""
-        else:
-            former = "a former "
 
+        # Rank icons from url.
         rank_name = row_values[10]
-        if rank_name == "Immortal" and current_member:
-            rank_url = "https://i.imgur.com/dpugisO.png"
-        elif rank_name == "Legendary" and current_member:
-            rank_url = "https://i.imgur.com/59Jighv.png"
-        elif rank_name == "Diamond" and current_member:
-            rank_url = "https://i.imgur.com/AZYAK39.png"
-        elif rank_name == "Gold" and current_member:
-            rank_url = "https://i.imgur.com/ZDLUlqs.png"
-        elif rank_name == "Silver" and current_member:
-            rank_url = "https://i.imgur.com/xxxlPAq.png"
-        elif rank_name == "Bronze" and current_member:
-            rank_url = "https://i.imgur.com/svAUm00.png"
-        else:
-            rank_url = "https://i.imgur.com/ys2UBNW.png"
+        rank_icons = {
+            "Immortal": "https://i.imgur.com/dpugisO.png",
+            "Legendary": "https://i.imgur.com/59Jighv.png",
+            "Diamond": "https://i.imgur.com/AZYAK39.png",
+            "Gold": "https://i.imgur.com/ZDLUlqs.png",
+            "Silver": "https://i.imgur.com/xxxlPAq.png",
+            "Bronze": "https://i.imgur.com/svAUm00.png",
+            "No rank": "https://i.imgur.com/ys2UBNW.png",
+        }
 
-        guild = member.guild
-        senior_tester_candidate_role = discord.utils.get(
-            guild.roles, name="Senior Tester Candidate"
-        )
-        senior_tester_role = discord.utils.get(guild.roles, name="Senior Tester")
-        staff_role = discord.utils.get(guild.roles, name="Frostburn Staff")
-        role_list = member.roles
-        if staff_role in role_list:
-            effective_role = "Frostburn staff member"
-            pass
-        elif senior_tester_role in role_list:
-            effective_role = "senior tester"
-        elif senior_tester_candidate_role in role_list:
-            effective_role = "senior tester candidate"
-        else:
-            effective_role = "tester"
+        role_translations = {
+            "None": "a former tester",
+            "Tester": "tester",
+            "Senior": "senior tester",
+            "Staff": "Frostburn staff member",
+        }
 
         nick_src = nick
-        if nick.startswith("`") or nick.startswith("_"):
-            nick = "\\" + nick
+        nick = discord.utils.escape_markdown(nick)
+        if current_member:
+            # Reusing leaderboard to calculate current placement.
+            offset = 2  # The first 2 rows are reserved.
+
+            # This testing cycle.
+            tc_players = [
+                list_of_lists[x][1]
+                for x in range(offset, len(list_of_lists))
+                if list_of_lists[x][0] != "None"
+            ]
+            tc_games = [
+                int(list_of_lists[x][2])
+                for x in range(offset, len(list_of_lists))
+                if list_of_lists[x][0] != "None"
+            ]
+            tc_bugs = [
+                int(list_of_lists[x][4])
+                for x in range(offset, len(list_of_lists))
+                if list_of_lists[x][0] != "None"
+            ]
+            tc_tokens = [
+                int(list_of_lists[x][8])
+                for x in range(offset, len(list_of_lists))
+                if list_of_lists[x][0] != "None"
+            ]
+
+            # Total games (all-time).
+            at_games = [
+                int(list_of_lists[x][5])
+                for x in range(offset, len(list_of_lists))
+                if list_of_lists[x][0] != "None"
+            ]
+            at_bugs = [
+                int(list_of_lists[x][7])
+                for x in range(offset, len(list_of_lists))
+                if list_of_lists[x][0] != "None"
+            ]
+
+            leaderboard_len = len(tc_players)
+
+            # Create a list of tuples and sort it later.
+            unsorted_leaderboard = []
+
+            for i in range(leaderboard_len):
+                unsorted_leaderboard.append(
+                    (
+                        tc_players[i].lower(),
+                        tc_games[i],
+                        tc_bugs[i],
+                        tc_tokens[i],
+                        at_games[i],
+                        at_bugs[i],
+                    )
+                )
+
+            sorted_by_games = sorted(
+                unsorted_leaderboard,
+                key=lambda unsorted_tuples: unsorted_tuples[1],
+                reverse=True,
+            )
+            sorted_by_bugs = sorted(
+                unsorted_leaderboard,
+                key=lambda unsorted_tuples: unsorted_tuples[2],
+                reverse=True,
+            )
+            sorted_by_tokens = sorted(
+                unsorted_leaderboard,
+                key=lambda unsorted_tuples: unsorted_tuples[3],
+                reverse=True,
+            )
+
+            sorted_by_at_games = sorted(
+                unsorted_leaderboard,
+                key=lambda unsorted_tuples: unsorted_tuples[4],
+                reverse=True,
+            )
+            sorted_by_at_bugs = sorted(
+                unsorted_leaderboard,
+                key=lambda unsorted_tuples: unsorted_tuples[5],
+                reverse=True,
+            )
+
+            # Convert cardinal to ordinal.
+            def ordinal(n):
+                _suffix = (
+                    "th"
+                    if 4 <= n % 100 <= 20
+                    else {1: "st", 2: "nd", 3: "rd"}.get(n % 10, "th")
+                )
+                return f"{n}{_suffix}"
+
+            player_tuple = (
+                nick_src.lower(),
+                int(games),
+                int(row_values[4]),
+                int(row_values[8]),
+                int(row_values[5]),
+                int(row_values[7]),
+            )
+
+            # Find player's placement on the ladder.
+            rank_tc_games = f"{ordinal(sorted_by_games.index(player_tuple)+1)}"
+            rank_tc_bugs = f"{ordinal(sorted_by_bugs.index(player_tuple)+1)}"
+            rank_tc_tokens = f"{ordinal(sorted_by_tokens.index(player_tuple)+1)}"
+            rank_at_games = f"{ordinal(sorted_by_at_games.index(player_tuple)+1)}"
+            rank_at_bugs = f"{ordinal(sorted_by_at_bugs.index(player_tuple)+1)}"
+
+        # Check if the player owns a Custom Account Icon on the retail client and display it.
+        _account_id = row_values[33]
+        if _account_id != "":
+            async with aiohttp.ClientSession() as session:
+                async with session.get(
+                    f"https://hon-avatar.now.sh/{_account_id}"
+                ) as resp:
+                    account_icon_url = await resp.text()
+                    if not account_icon_url.endswith(".cai"):
+                        account_icon_url = "https://s3.amazonaws.com/naeu-icb2/icons/default/account/default.png"
+        else:
+            account_icon_url = (
+                "https://s3.amazonaws.com/naeu-icb2/icons/default/account/default.png"
+            )
 
         embed = discord.Embed(
             title="Retail Candidate Testers",
             type="rich",
-            description="Information for {0}{1} {2}.".format(
-                former, effective_role, nick
-            ),
+            description=f"Information for {role_translations[row_values[0]]} {nick}.",
             url="https://forums.heroesofnewerth.com/forumdisplay.php?209-Retail-Candidate-Testers",
             color=0xFF6600,
-            timestamp=ctx.message.created_at,
+            timestamp=config.LAST_RETRIEVED,
         )
         embed.set_author(
             name=nick_src,
-            url="https://docs.google.com/spreadsheets/d/1HpDhrKbyK01rRUUFw3v-30QrNJpLHaOmWYrUxkKDbq0/edit#gid=0",
-            icon_url=member.avatar_url,
+            url=f"https://forums.heroesofnewerth.com/member.php?{_account_id}",
+            icon_url=account_icon_url,
         )
         if current_member:
             embed.add_field(name="Unconfirmed games", value=games_played, inline=True)
             # embed.add_field(name=u"\u2063", value=u"\u2063", inline=True)
-            embed.add_field(name="Games", value=games, inline=True)
-        embed.add_field(name="Total games", value=row_values[5], inline=True)
+            embed.add_field(
+                name="Games", value=f"{games} ({rank_tc_games})", inline=True
+            )
+        if current_member:
+            embed.add_field(
+                name="Total games",
+                value=f"{row_values[5]} ({rank_at_games})",
+                inline=True,
+            )
+        else:
+            embed.add_field(name="Total games", value=row_values[5], inline=True)
         if current_member:
             embed.add_field(name="Game time", value=gametime, inline=True)
         embed.add_field(name="Total game time", value=gametime_total, inline=True)
         if current_member:
-            embed.add_field(name="Bug reports", value=row_values[4], inline=True)
-        embed.add_field(name="Total bug reports", value=row_values[7], inline=True)
-        # embed.add_field(name="Total games", value=row_values[3], inline=True)
-        # embed.add_field(name="Total game time", value="N/A", inline=True)
-        # embed.add_field(name="Total bug reports", value=row_values[10], inline=True)
-        if current_member:
-            embed.add_field(name="Tokens earned", value=row_values[8], inline=True)
-            embed.add_field(name="Bonuses", value=bonus, inline=True)
-            embed.add_field(name="Activity rank", value=row_values[10], inline=True)
             embed.add_field(
-                name="Multiplier", value="{0}x".format(row_values[12]), inline=True
+                name="Bug reports",
+                value=f"{row_values[4]} ({rank_tc_bugs})",
+                inline=True,
             )
+        if current_member:
+            embed.add_field(
+                name="Total bug reports",
+                value=f"{row_values[7]} ({rank_at_bugs})",
+                inline=True,
+            )
+        else:
+            embed.add_field(
+                name="Total bug reports", value=row_values[7], inline=True,
+            )
+        if current_member:
+            embed.add_field(
+                name="Tokens earned",
+                value=f"{row_values[8]} ({rank_tc_tokens})",
+                inline=True,
+            )
+            embed.add_field(name="Bonuses", value=bonus, inline=True)
+            embed.add_field(name="Activity rank", value=rank_name, inline=True)
+            embed.add_field(name="Multiplier", value=f"{row_values[12]}x", inline=True)
             embed.add_field(name="Perks", value=row_values[19], inline=True)
             embed.add_field(name="Absence", value=absence, inline=True)
         embed.add_field(name="Join date", value=row_values[20], inline=True)
@@ -235,56 +358,47 @@ class Stats(commands.Cog):
             embed.add_field(
                 name="Reason for removal", value=row_values[22], inline=False
             )
-        # embed.add_field(name=u"\u2063", value=u"\u2063", inline=True)
-        # embed.add_field(name=u"\u2063", value=u"\u2063", inline=True)
         if row_values[31] != "":
-            embed.add_field(
-                name="Awards", value=u"\u2063" + row_values[31], inline=True
-            )
+            embed.add_field(name="Awards", value="\u2063" + row_values[31], inline=True)
         if requester_discord_id is not None:
             embed.set_footer(
-                text="Requested by {0} (✓). React with 🆗 to delete this message.".format(
-                    requester_name
-                ),
+                text=f"Requested by {requester_name} (✓). Timestamp provided shows the time of last retrieval. React with 🗑️ to delete or with 💾 to preserve this message. No action results in deletion after 5 minutes.",
                 icon_url="https://i.imgur.com/q8KmQtw.png",
             )
         else:
             embed.set_footer(
-                text="Requested by {0} ({1}#{2}). React with 🆗 to delete this message.".format(
-                    ctx.message.author.display_name,
-                    ctx.message.author.name,
-                    ctx.message.author.discriminator,
+                text="Requested by {.display_name} ({.name}#{.discriminator}). Timestamp provided shows the time of last retrieval. React with 🗑️ to delete or with 💾 to preserve this message. No action results in deletion after 5 minutes.".format(
+                    ctx.author
                 ),
                 icon_url="https://i.imgur.com/q8KmQtw.png",
             )
         # embed.set_footer(text="React with 🆗 to delete this message.", icon_url="https://i.imgur.com/Ou1k4lD.png")
         # embed.set_thumbnail(url="https://i.imgur.com/q8KmQtw.png")
-        embed.set_thumbnail(url=rank_url)
+        embed.set_thumbnail(url=rank_icons[rank_name])
         message = await ctx.send(embed=embed)
         stop = timeit.default_timer()
         print(stop - start)
-        await message.add_reaction("🆗")
+        await message.add_reaction("🗑️")
+        await message.add_reaction("💾")
+
         try:
-            await self.bot.wait_for(
+            reaction, _ = await self.bot.wait_for(
                 "reaction_add",
                 check=lambda reaction, user: user == ctx.message.author
-                and reaction.emoji == "🆗"
+                and reaction.emoji in ["🗑️", "💾"]
                 and reaction.message.id == message.id,
-                timeout=120.0,
+                timeout=300.0,
             )
-            await message.delete()
+
+            if reaction.emoji == "🗑️":
+                await message.delete()
+
         except asyncio.TimeoutError:
             await message.delete()
 
     # TO DO: this needs to be prettier
     @commands.command(aliases=["hero", "herousage"])
-    @commands.has_any_role(
-        "Manage Roles",
-        "Overlord",
-        "Frostburn Staff",
-        "Senior Tester",
-        "Senior Tester Candidate",
-    )  # TO DO: alternative
+    @is_senior()
     @database_ready()
     async def usage(self, ctx):
         """Hero usage list for the current patch cycle."""
